requirements.txt
# Deep Learning
tensorflow>=2.10.0,<2.18.0
keras>=2.10.0

# Data Processing
numpy>=1.24.3,<2.0.0
pandas>=1.5.3,<3.0.0

# Visualization
matplotlib>=3.5.2,<4.0.0

# Image Processing
pillow>=9.2.0,<11.0.0

# Machine Learning
scikit-learn>=1.1.3,<2.0.0

# Development
jupyter>=1.0.0
python-dotenv>=0.20.0